refactor: add tests and database to common area
- move test and database code under watsononlinestore/
- move fakes to tests
- add a unit module under tests to better represent it
- fixed up imports to be more openstack-ish
  (alphabetical, and grouped by: stdlib, 3rd party, local)
