+++ watsononlinestore/tests/unit/test_watson_online_store.py
@@ -1,7 +1,9 @@
-from watsononlinestore import watson_online_store
+import unittest
+
 import ddt
 import mock
-import unittest
+
+from watsononlinestore import watson_online_store
 
 
 @ddt.ddt
